==> 1-ETL_JSON(output_steam_games).ipynb <==
# Extracción, Transformación y Carga (ETL) del archivo "output_steam_games.json" 

## Importamos librerías 

import json
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt 
import seaborn as sns
import pyarrow as pa
import pyarrow.parquet as pq

## Extraemos y exploramos la data 

# Leemos el archivo json y lo guardamos en una lista
row = [] # creamos una lista vacia para ir agregando las filas del archivo json
with open ("Datasets/output_steam_games.json") as file: # abrimos el archivo json
    for line in file.readlines(): # leemos linea por linea
        data = json.loads(line)
        row.append(data)

# Convertimos la lista en un dataframe 
games = pd.DataFrame(row)
games.head()

games.shape # Obtenemos la cantidad de filas y columnas

games.describe() # Obtenemos un resumen de las columnas numéricas

games.info() # Obtenemos informacion de las columnas

# revisamos el porcentaje de filas completamente nulas respecto al total de filas del dataframe
total_filas = len(games)
filas_nulas = games.isnull().all(axis=1).sum()
porcentaje_filas_nulas = (filas_nulas / total_filas) * 100
porcentaje_filas_nulas

# verificamos el porcentaje de nulos por columna para tomar decisiones
total_filas = len(games)
nulos_por_columna = games.isnull().sum()
porcentaje_nulos_por_columna = (nulos_por_columna / total_filas) * 100
porcentaje_nulos_por_columna

**Observación**
- La cantidad de filas con valores completamente nulos representa un 73 % del total del dataframe. 
- Todas las columnas cuentan con mas del 70 % de valores nulos.

### Verificamos cantidad de datos nulos y duplicados

nulos_por_columna = games.isnull().sum() # Obtenemos la cantidad de nulos por columna
nulos_por_columna

total_nulos = games.isnull().sum().sum() # Obtenemos la cantidad total de nulos
total_nulos

steam_games = games.dropna(how="all").reset_index(drop=True) # Eliminamos las filas que tienen "solo valores nulos"
steam_games.head()

total_nulos = steam_games.isnull().sum().sum() # Verificamos nuevamente la cantidad total de nulos
total_nulos

**Observación**
- Al eliminar las filas que solo tienen valores nulos, disminuimos la cantidad de los mismos de 1168997 a 20967.

duplicados_columnas = steam_games[steam_games.duplicated(subset=['id'])] # Verificamos si hay filas duplicadas en la columna id
duplicados_columnas

steam_games[steam_games["developer"] == "Machine Games"]

#### Las filas duplicadas 14573 y 13894 son iguales por lo que podemos eliminar una de ellas.
#### La fila 30961 está duplicada y es idéntica a la fila 1068, pero contiene un valor nulo en el campo de identificación (ID), mientras que la fila 1068 tiene un ID completo de 200260. En consecuencia, la fila duplicada 30961 se elimina.

steam_games = steam_games.drop([14573, 30961]) # Eliminamos columnas con id duplicadas

duplicados_columnas = steam_games[steam_games.duplicated(subset=['id'])]
duplicados_columnas

steam_games.shape

# verificamos el porcentaje de nulos por columna para tomar una decision
total_filas = len(steam_games)
nulos_por_columna = steam_games.isnull().sum()
porcentaje_nulos_por_columna = (nulos_por_columna / total_filas) * 100
porcentaje_nulos_por_columna

**Observación**
- La columna 'publisher' cuenta con un 25% de datos nulos siendo la de mayor cantidad.

## Transformación de los datos 

tipos_de_datos = steam_games.dtypes # Obtenemos los tipos de datos de cada columna
tipos_de_datos

#### El conjunto de datos sigue siendo un archivo JSON, por lo que el tipo de dato es 'object', lo que podría indicar datos anidados con formato clave-valor. Se procede a analizar cada columna mediante un bucle, almacenando la información en un diccionario vacío. Posteriormente, se transforma este diccionario en un conjunto de datos para poder examinar los tipos de datos de las columnas.

tipo_data = {"columna":[],"tipos_de_datos":[]} # creamos un diccionario vacio

for columna in steam_games.columns: # iteramos por las columnas del dataframe 
    tipo_data["columna"].append(columna) # agregamos el nombre de la columna
    tipo_data["tipos_de_datos"].append(steam_games[columna].apply(type).unique()) # agregamos el tipo de dato de la columna

analisis= pd.DataFrame(tipo_data) # convertimos el diccionario en un dataframe
analisis

### Columna "release_date"

#### La columna "release_date" contiene la fecha de lanzamiento de cada juego de forma YYYY-MM-DD. Transformamos esta columna de manera que solo veamos el año de lanzamiento.

# observamos la composición de la columna
steam_games["release_date"]

steam_games["release_date"].value_counts() # observamos la cantidad de valores por fecha

steam_games["release_date"].isnull().sum() # observamos la cantidad de nulos

# Convertimos la columna en datetime
steam_games['release_date'] = pd.to_datetime(steam_games['release_date'], errors='coerce')

# Calculamos la fecha media
mean_date = steam_games['release_date'].mean()

# Rellenamos los valores nulos con la fecha media
steam_games['release_date'].fillna(mean_date, inplace=True)

# Verificamos que no hayan nulos
steam_games["release_date"].isnull().sum()

# Extraemos el año de la columna "release_date" y creamos una nueva columna llamada "release_year"
steam_games['release_year'] = steam_games['release_date'].dt.year 
steam_games.head(2)

#### Al contar con la columna 'release_year' optamos por eliminar 'release_date'

steam_games = steam_games.drop('release_date', axis=1)

### Columna "price"

steam_games["price"] # observamos la composición de la columna

steam_games["price"].value_counts() # observamos la cantidad de valores por precio

# Definimos una función para poder reemplazar los datos string de la columna "price" por flotantes 0.0

def str_a_float(valor):

    if pd.isna(valor): # si el valor es nulo retorno 0.0
        return 0.0
    try:
        flotante = float(valor) # si el valor es un flotante retorno el mismo valor
        return flotante
    except (ValueError, TypeError): # si el valor no es un flotante retorno 0.0
        return 0.0


steam_games["price"] = steam_games["price"].apply(str_a_float) # aplicamos la función a la columna "price"

steam_games["price"] # observamos la nueva composición de la columna

### Columna "genres"

steam_games["genres"][3] # observamos la composición de un campo al azar y vemos que es una lista

steam_games["genres"].isnull().sum() # verificamos la cantidad de nulos

steam_games["genres"].value_counts() # verificamos la cantidad de valores por genero

#### Aplicamos una función para limpiar la columna 'genres' y rellenar las listas con el string de índice 0 de cada lista y los valores nulos con "Pending classification"

def clean_genres(genres):
    if isinstance(genres, list):
        return genres[0]
    elif isinstance(genres, str):
        return genres
    else:
        return 'Pending classification'

# Aplicamos la función a la columna 'genres' y creamos una nueva columna 'genres_clean'
steam_games['genres_clean'] = steam_games['genres'].apply(clean_genres)

steam_games["genres_clean"].value_counts() # verificamos la cantidad de valores por genero

#### Observamos que 'Pending classification' ocupa el quinto lugar entre los géneros. Por esta razón, decidimos implementar una clase que contenga funciones para los géneros más conocidos. Estas funciones se encargarán de recorrer la nueva columna 'genres_clean'. En caso de encontrar 'Pending classification' en algún campo, reemplazarán ese valor con el género correspondiente encontrado en la columna 'tags'. Este enfoque nos permite evitar la pérdida de información.

class GenreClassifier:
    def replace_pending_with_action(self, row):
        if row['genres_clean'] == 'Pending classification':
            if isinstance(row['tags'], list):
                if 'Action' in row['tags']:
                    return 'Action'
        return row['genres_clean']

    def replace_pending_with_indie(self, row):
        if row['genres_clean'] == 'Pending classification':
            if isinstance(row['tags'], list):
                if 'Indie' in row['tags']:
                    return 'Indie'
        return row['genres_clean']

    def replace_pending_with_casual(self, row):
        if row['genres_clean'] == 'Pending classification':
            if isinstance(row['tags'], list):
                if 'Casual' in row['tags']:
                    return 'Casual'
        return row['genres_clean']

    def replace_pending_with_adventure(self, row):
        if row['genres_clean'] == 'Pending classification':
            if isinstance(row['tags'], list):
                if 'Adventure' in row['tags']:
                    return 'Adventure'
        return row['genres_clean']

    def replace_pending_with_strategy(self, row):
        if row['genres_clean'] == 'Pending classification':
            if isinstance(row['tags'], list):
                if 'Strategy' in row['tags']:
                    return 'Strategy'
        return row['genres_clean']

    def replace_pending_with_rpg(self, row):
        if row['genres_clean'] == 'Pending classification':
            if isinstance(row['tags'], list):
                if 'RPG' in row['tags']:
                    return 'RPG'
        return row['genres_clean']

    def replace_pending_with_simulation(self, row):
        if row['genres_clean'] == 'Pending classification':
            if isinstance(row['tags'], list):
                if 'Simulation' in row['tags']:
                    return 'Simulation'
        return row['genres_clean']

    def replace_pending_with_racing(self, row):
        if row['genres_clean'] == 'Pending classification':
            if isinstance(row['tags'], list):
                if 'Racing' in row['tags']:
                    return 'Racing'
        return row['genres_clean']

    def replace_pending_with_free_to_play(self, row):
        if row['genres_clean'] == 'Pending classification':
            if isinstance(row['tags'], list):
                if 'Free to Play' in row['tags']:
                    return 'Free to Play'
        return row['genres_clean']

# Creamos una instancia de la clase
genre_classifier = GenreClassifier()

# Aplicamos las funciones a la columna 'genres_clean'
steam_games['genres_clean'] = steam_games.apply(genre_classifier.replace_pending_with_action, axis=1)

steam_games['genres_clean'] = steam_games.apply(genre_classifier.replace_pending_with_indie, axis=1)

steam_games['genres_clean'] = steam_games.apply(genre_classifier.replace_pending_with_casual, axis=1)

steam_games['genres_clean'] = steam_games.apply(genre_classifier.replace_pending_with_adventure, axis=1)

steam_games['genres_clean'] = steam_games.apply(genre_classifier.replace_pending_with_strategy, axis=1)

steam_games['genres_clean'] = steam_games.apply(genre_classifier.replace_pending_with_rpg, axis=1)

steam_games['genres_clean'] = steam_games.apply(genre_classifier.replace_pending_with_simulation, axis=1)

steam_games['genres_clean'] = steam_games.apply(genre_classifier.replace_pending_with_free_to_play, axis=1)

steam_games["genres_clean"].unique()

steam_games["genres_clean"].value_counts()

steam_games["genres"].isnull().sum() # verificamos la cantidad de nulos de la columna "genres"

steam_games["genres_clean"].isnull().sum() # verificamos la cantidad de nulos de la columna "genres_clean"

#### Al comparar la cantidad de nulos de las columnas 'genres y 'genres_clean', tomamos las decisión de eliminar la columna 'genres' y quedarnos con la columna 'genres_clean' que cuenta con un 100% de datos no nulos.

# Eliminamos la columna 'genres'
steam_games = steam_games.drop('genres', axis=1)

# Renombramos 'genres_clean' como 'genres'
steam_games = steam_games.rename(columns={'genres_clean': 'genres'})

steam_games.info() # obtenemos información de las columnas en búsqueda de nulos

total_filas = len(steam_games)
nulos_por_columna = steam_games.isnull().sum()
porcentaje_nulos_por_columna = (nulos_por_columna / total_filas) * 100
porcentaje_nulos_por_columna

#### Por último y para finalizar el trabajo en esta columna y ya habiendo cumplido con su propósito, eliminamos la columna 'tags'.

#Eliminamos la columna 'tags'
steam_games = steam_games.drop('tags', axis=1)

### Columna 'specs'

#### Esta columna, almacena información sobre especificaciones técnicas de los juegos, tales como si es "Single-player", "Multi-player", "Co-op", etc., haciendo que la misma no sea de mucha utilidad para nuestro análisis. Por esta razón, decidimos eliminarla.

steam_games = steam_games.drop('specs', axis=1)

### Columnas "publisher" y "developer"

#### Apreciamos que las columnas 'publisher' y 'developer' cuentan con un 25% y 10% de datos nulos respectivamente. Por esta razón, y en función de que suelen estar directamente relacionadas, optamos por aplicar una funcion que nos permita rellenar los valores nulos de la columna 'developer' con los valores coincidentes en el mismo campo de la columna 'publisher'.

def fill_null_developers_with_publisher(row):
    if pd.isnull(row['developer']):
        return row['publisher']
    return row['developer']

# AplicaMOS la función a la columna 'developer'
steam_games['developer'] = steam_games.apply(fill_null_developers_with_publisher, axis=1)

steam_games.info()

#### Teniendo en cuenta que las consignas de funciones nos piden trabajar con 'DESARROLLADORES', optamos por eliminar la columna 'publisher' y quedarnos con la columna 'developer'.

# Eliminamos la columna 'publisher'
steam_games = steam_games.drop('publisher', axis=1)

#### Basándonos en un análisis previo y para evitar el sesgo que podría generar rellenar los campos nulos de la columna 'developer' con 'No Data', siendo esta columna de mucha importancia para los endpoints. Decidimos eliminar las filas que contengan valores nulos en la columna 'developer'.

# Filtramos y mantenemos solo las filas donde 'developer' no sea nulo
steam_games = steam_games[steam_games['developer'].notnull()]

# Reestablecemos los índices despúes de filtrar
steam_games = steam_games.reset_index(drop=True)

### Columna "id"

# Rellenamos los valores NaN en la columna 'id' con 'No Data'
steam_games['id'] = steam_games['id'].fillna('No Data')

# Convertimos la columna 'id' a tipo string
steam_games['id'] = steam_games['id'].astype(str)

# Renombramos la columna 'id' como 'item_id'
steam_games = steam_games.rename(columns={'id': 'item_id'})

# Verificamos la información de las columnas
steam_games.info()

### Columnas "early_access", "url" y "review_url"

#### Estas columnas no aportan información relevante para nuestro análisis, por lo que decidimos eliminarlas.

# Eliminamos las columnas 'early_access', 'url' y 'reviews_url'
steam_games = steam_games.drop(['early_access', 'url', 'reviews_url'], axis=1)

### Columnas "app_name" y "title"

#### Las columnas 'app_name' y 'title' cuentan con un 0.006% y 6.37% de datos nulos respectivamente. Por esta razón, y en función de que estan directamente relacionadas, optamos por eliminar la segunda.

# Eliminamos la columna 'title'
steam_games = steam_games.drop('title', axis=1)
steam_games.info()

## Guardamos el dataset como CSV

steam_games.to_csv("Datasets/output_steam_games.csv", index=False, encoding="utf-8")

#### Convertimos el archivo CSV a formato Parquet con el objetivo de optimizar el almacenamiento y mejorar la eficiencia en las operaciones de lectura y escritura. La elección de transformación a Parquet se debe a la facilidad de trabajo con conjuntos de datos complejos. Esta conversión facilita el procesamiento distribuido y ofrece opciones avanzadas de compresión, contribuyendo así a una gestión más eficiente de los recursos y a un mejor rendimiento en entornos de procesamiento de datos.

steam_games = pd.read_csv("Datasets/output_steam_games.csv") # leemos el archivo csv

tabla = pa.Table.from_pandas(steam_games) # convertimos el dataframe en una tabla
pq.write_table(tabla,"Datasets/output_steam_games.parquet") # guardamos la tabla en un archivo parquet
