==> 2-ETL_JSON(australian_users_items).ipynb <==
# Extracción, Transformación y Carga (ETL) del archivo "australian_users_items.json"

## Importamos librerías

import json
import ast
import re
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt 
import seaborn as sns
import pyarrow as pa
import pyarrow.parquet as pq

## Extraemos y exploramos la data

row = []  # creamos una lista vacia para ir agregando las filas del archivo json

with open("Datasets/australian_users_items.json", encoding="utf-8") as file:
    for line in file:
        # Utilizamos ast.literal_eval para interpretar la línea como expresión literal de Python
        # Luego, convertimos la expresión evaluada a formato JSON utilizando json.dumps
        row.append(json.loads(json.dumps(ast.literal_eval(line))))

# Creamos un dataframe con la lista de diccionarios
items = pd.DataFrame(row)
items

items.shape # vemos la cantidad de filas y columnas

## Transformamos el dataset

#### La columna "items" alberga datos anidados en forma de una lista de diccionarios. En este conjunto de datos, se utiliza el método de normalización, donde la columna "items" sirve como columna organizadora. De esta manera, se obtienen las claves de los diccionarios como columnas individuales.

# normalizamos la columna 'items'
normalizado = pd.json_normalize(row, record_path=['items'], meta=['steam_id','items_count','user_id', 'user_url'] )
normalizado

tipo_data = {"columna":[],"tipos_de_datos":[]} # generamos un diccionario vacio para ir almacenando lo que genere el bucle

for columna in normalizado.columns: # recorremos las columnas del dataframe 
    tipo_data["columna"].append(columna) # agregamos el nombre de la columna
    tipo_data["tipos_de_datos"].append(normalizado[columna].apply(type).unique()) # agregamos los tipos de datos que hay en la columna

analisis= pd.DataFrame(tipo_data) # creamos un dataframe con el diccionario
analisis

normalizado.info() # vemos los tipos de datos de cada columna

#### En este dataframe, la búsqueda y eliminación de valores nulos se llevará a cabo después de la normalización. Esta decisión se basa en la consideración de la abundante información contenida en la columna anidada "items". La intención es evitar la pérdida de datos sin previamente haber realizado un análisis exhaustivo de los mismos.

duplicados = normalizado.loc[normalizado.duplicated()] # buscamos los duplicados
duplicados

### Analizamos las filas con datos duplicados y se eliminan

normalizado = normalizado.drop_duplicates(keep='first')
normalizado

#### Teniendo en cuenta que este conjunto de datos contiene información de usuarios, como por ejemplo las horas de juego, y viendo que lo que necesitamos para los endpoints esta directamente relacionado con ello y que a su vez contamos con mas de 5.000.000 de registros; lo cual hace que los tiempos de carga sean superiores a otros datasets, revisaremos la cantidad de filas donde las horas de juego sean 0 y si vemos un número importante de registros, los eliminaremos.

normalizado['playtime_forever'].value_counts()

# Filtramos y mantenemos solo las filas donde playtime_forever no es igual a 0
normalizado = normalizado[normalizado['playtime_forever'] != 0]

# Restablecemos los índices después de filtrar
normalizado = normalizado.reset_index(drop=True)

normalizado['playtime_forever'].value_counts()

#### Teniendo en cuenta que en la columna playtime_forever los datos se encuentran en minutos, se procede a convertirlos a horas, creando una nueva columna para tal fin.

# Creamos la columna 'playtime_forever_hours' que es el resultado de dividir 'playtime_forever' por 60 y redondeando a 2 decimales
normalizado['playtime_hours'] = round(normalizado['playtime_forever']/60,2)

normalizado['playtime_hours'].value_counts() # vemos la cantidad de horas jugadas

#### Siguiendo por el mismo camino, y al ver que las columnas 'playtime_forever' y 'playtime_2weeks' se hacen innecesarias, eliminaremos las mismas para reducir aun mas el tamaño del dataset.

# Eliminamos las columnas 'playtime_2weeks' y 'playtime_2weeks' por no ser de utilidad para el análisis
normalizado = normalizado.drop(['playtime_2weeks','playtime_forever'], axis=1)

normalizado.info()

### Columna 'user_url'

#### La columna 'user_url' contiene la url del usuario en la plataforma Steam. Esta columna no es necesaria para el análisis y se eliminará.

# Eliminamos la columna 'usr_url'
normalizado = normalizado.drop(['user_url'], axis=1)

## Al igual que en el ETL anterior, guardamos el dataframe en formato parquet para optimizar el almacenamiento y mejorar la eficiencia en las operaciones de lectura y escritura

normalizado = normalizado.copy() # copiamos el dataframe para no modificar el original

normalizado.to_csv("Datasets/australian_users_items.csv", index=False, encoding="utf-8") # guardamos el dataframe en un archivo csv

normalizado = pd.read_csv("Datasets/australian_users_items.csv") # leemos el archivo csv

tabla = pa.Table.from_pandas(normalizado) # convertimos el dataframe en una tabla de pyarrow
pq.write_table(tabla,"Datasets/australian_users_items.parquet") # guardamos la tabla en un archivo parquet
