==> 5-Funciones_Endpoints.ipynb <==
# Creación de funciones para endpoints

## Importamos librerías

import pandas as pd
import pyarrow as pa
import pyarrow.parquet as pq
import warnings
warnings.filterwarnings('ignore')

df_games = pd.read_parquet("Datasets/output_steam_games.parquet") # Cargamos el dataset de juegos
df_games.head(2)

df_games.info()

df_user_items = pd.read_parquet("Datasets/australian_users_items.parquet") # Cargamos el dataset de items
df_user_items.head(2)

df_reviews = pd.read_parquet("Datasets/australian_user_reviews.parquet") # Cargamos el dataset de reviews
df_reviews.head(2)

df_reviews.info()

df_games_and_reviews = pd.read_parquet("Datasets/games_and_reviews.parquet") # Cargamos el dataset de reviews
df_games_and_reviews.head(2)

df_reviews

### Función `PlayTimeGenre`

#### Función que devuelve el año con mas horas de juego para un género

def PlayTimeGenre(genero, df_games: pd.DataFrame, df_user_items: pd.DataFrame):
    # Convertimos el género proporcionado a minúsculas para una comparación sin distinción de mayúsculas
    genero = genero.lower()
    
    # Filtramos los juegos por género de manera más flexible
    df_filtered = df_games[df_games['genres'].str.lower().str.contains(fr'\b{genero}\b', na=False)]

    if not df_filtered.empty:
        # Convertimos 'item_id' a tipo de dato object en ambos DataFrames
        df_user_items['item_id'] = df_user_items['item_id'].astype(str)
        df_filtered['item_id'] = df_filtered['item_id'].astype(str)

        df_merged = pd.merge(df_user_items, df_filtered[['item_id', 'release_year']], left_on='item_id', right_on='item_id')
        
        # Verificamos si la longitud de df_filtered es mayor a cero, es decir, se encontraron géneros
        result = {"Año con más horas jugadas para el género {}: {}".format(genero.capitalize(), df_merged.groupby('release_year')['playtime_hours'].sum().idxmax())}
    else:
        result = {"Año con más horas jugadas para género {}: {}".format(genero.capitalize(), "Género no encontrado en la base de datos")}

    return result

PlayTimeGenre("Sports", df_games, df_user_items)

### Función `UserForGenre`

#### Función que devuelve el usuario con más horas jugadas para un género 

def UserForGenre(genero: str, df_games: pd.DataFrame, df_user_items: pd.DataFrame):
    # Convertimos el género proporcionado a minúsculas para una comparación sin distinción de mayúsculas
    genero = genero.lower()
    
    # Filtramos los juegos por género de manera más flexible
    df_filtered = df_games[df_games['genres'].str.lower().str.contains(fr'\b{genero}\b', na=False)]
    
    if not df_filtered.empty:
        df_user_items['item_id'] = df_user_items['item_id'].astype(str)
        
        # Convertimos la columna 'item_id' en df_filtered a tipo 'str'
        df_filtered['item_id'] = df_filtered['item_id'].astype(str)
        
        df_merged = pd.merge(df_user_items, df_filtered[['item_id', 'release_year']], on='item_id')
        
        # Filtramos los datos desde el año 2003 en adelante
        df_merged = df_merged[df_merged['release_year'] >= 2003]
        user_with_most_playtime = df_merged.groupby('user_id')['playtime_hours'].sum().idxmax()
        playtime_by_year = df_merged.groupby(['release_year', 'user_id'])['playtime_hours'].sum().reset_index()
        playtime_by_year = playtime_by_year[playtime_by_year['user_id'] == user_with_most_playtime]
        playtime_by_year = playtime_by_year.rename(columns={'release_year': 'Año', 'playtime_hours': 'Horas'})
        
        # Convertimos las horas a enteros
        playtime_by_year['Horas'] = playtime_by_year['Horas'].astype(int)
        
        result = {
            "Usuario con más horas jugadas para Género {}:".format(genero.capitalize()): user_with_most_playtime,
            "Horas jugadas": [{"Año": str(row['Año']), "Horas": row['Horas']} for _, row in playtime_by_year.iterrows()]
        }
    else:
        result = {"Usuario con más horas jugadas para Género {}:".format(genero.capitalize()): "Género no encontrado en la base de datos", "Horas jugadas": []}
    
    return result


UserForGenre("Action", df_games, df_user_items)

### Función `UsersRecommend`

#### Función que devuelve un Top 3 de los juegos mas recomendados por año

def UsersRecommend(año: int, df_games_and_reviews: pd.DataFrame):
    # Verificamos si el año está dentro del rango esperado
    rango_aceptado = range(2010, 2016)
    if año not in rango_aceptado:
        return {"message": "Mi base de datos solo tiene registros entre 2010 y 2015"}

    # Filtramos por el año deseado
    df_filtered = df_games_and_reviews[df_games_and_reviews['release_year'] == año]

    # Filtramos por comentarios recomendados y sentiment_analysis positivo/neutral
    df_filtered = df_filtered[(df_filtered['recommend'] == True) & (df_filtered['sentiment_analysis'].isin([0, 1, 2]))]

    # Obtenemos el top 3 de juegos recomendados
    top_games = df_filtered['app_name'].value_counts().head(3).reset_index()
    top_games = top_games.rename(columns={'index': 'Puesto 1', 'app_name': 'Juego'})

    # Modificamos la estructura del resultado
    result = [{"Puesto {}".format(i + 1): juego, 'recomendaciones': count} for i, (juego, count) in enumerate(zip(top_games['Juego'], top_games['count']))]

    return result

UsersRecommend(2014, df_games_and_reviews)

### Función `UsersWorstDeveloper`

#### Función que devuelve el Top 3 de peor desarrollador del año

def UsersWorstDeveloper(año: int, df_games_and_reviews: pd.DataFrame):
    # Verificamos si el año está dentro del rango esperado
    rango_aceptado = range(2010, 2016)
    if año not in rango_aceptado:
        return {"message": "Mi base de datos solo tiene registros entre 2010 y 2015"}

    # Filtramos por comentarios no recomendados y sentiment_analysis negativo
    df_filtered = df_games_and_reviews[(df_games_and_reviews['recommend'] == False) & (df_games_and_reviews['sentiment_analysis'] == 0)]

    # Filtramos por el año deseado
    df_filtered_year = df_filtered[df_filtered['release_year'] == año]

    # Si no hay datos para el año, retornamos mensaje
    if not df_filtered_year.empty:
        # Obtener los top 3 desarrolladores con menos recomendaciones
        top_developers = df_filtered_year['developer'].value_counts().head(3).reset_index()
        top_developers = top_developers.rename(columns={'index': 'Puesto 1', 'developer': 'Desarrollador'})

        # Modificamos la estructura del resultado
        result = [{"Puesto {}".format(i + 1): desarrollador} for i, desarrollador in enumerate(top_developers['Desarrollador'])]
    else:
        result = {"No hay juegos no recomendados para el año {}".format(año)}

    return result

UsersWorstDeveloper(2010, df_games_and_reviews)

### Función `sentiment_analysis`

#### Función que devuelve el conteo de las reviews positivas, neutrales y negativas de una empresa desarrolladora

def sentiment_analysis(empresa_desarrolladora: str, df_games_and_reviews: pd.DataFrame):
    # Filtramos por desarrolladora
    df_filtered_developer = df_games_and_reviews[df_games_and_reviews['developer'] == empresa_desarrolladora]

    # Verificamos que haya datos para la desarrolladora
    if not df_filtered_developer.empty:
        # Contamos los sentimientos y mapeamos el número del sentimiento a su etiqueta correspondiente
        sentiment_counts = df_filtered_developer['sentiment_analysis'].value_counts().reset_index(drop=True)
        sentiment_mapping = {0: 'Negative', 1: 'Neutral', 2: 'Positive'}
        sentiment_counts.index = sentiment_counts.index.map(sentiment_mapping)
        
        result = {empresa_desarrolladora: sentiment_counts.to_dict()}
    else:
        result = 'No cuento con los registros de esa empresa en mi base de datos'

    return result

sentiment_analysis('Treyarch', df_games_and_reviews)

# `Observación:` Al encontrar problemas con la ejecución de las funciones de endpoints en el archivo 'app.py', se optó por modificar la estructura de algunas para que puedan ser ejecutadas con FastAPI y ser deployadas.
