==> 6-Modelado_ML.ipynb <==
## Modelado de datos para Funciones de Machine Learning

### Importamos librerías 

import pandas as pd
import numpy as np
import scipy as sp
from sklearn.metrics.pairwise import cosine_similarity
import operator
import fastparquet as fp 
import pyarrow as pa
import pyarrow.parquet as pq


### Modelo de recomendación Item_Item

##### Un modelo de recomendación `item-item` basado en la similitud del coseno se configura para sugerir elementos o ítems que son similares a aquellos con los que un usuario ha interactuado o manifestado interés en el pasado. Este enfoque se sustenta en la premisa de que si a un usuario le agradó un determinado ítem, es probable que también disfrute de ítems con características similares.

##### El perfil de un ítem se representa comúnmente como un vector que incorpora información relevante sobre ese ítem en particular. Por ejemplo, en un sistema de recomendación de películas, el perfil de una película podría abarcar detalles sobre su género, director, elenco de actores, etiquetas asociadas, entre otros aspectos. Este enfoque permite al sistema discernir similitudes entre ítems y proporcionar recomendaciones más precisas, adaptadas a las preferencias específicas del usuario.

#### Iniciamos con la creación de dummies, para luego crear la matriz necesaria para el análisis de similitud de coseno

modelo_item = pd.read_parquet("Datasets/modelo_item.parquet")
modelo_item.head(3)

Generamos un DataFrame que incluye en sus columnas los géneros que pueden ser combinados con los nombres de los juegos en cada una de las filas. 

#Creamos los dummies
modelo_item = pd.get_dummies(modelo_item, columns=["genres"], prefix="")

modelo_item = modelo_item.groupby(["item_id","app_name"]).sum().reset_index()

modelo_item.head(4)

#### Calculamos la similitud del coseno

Procedemos a calcular la similitud del coseno entre todos los pares posibles de ítems basándose en sus perfiles. La similitud del coseno evalúa el ángulo entre dos vectores en un espacio multidimensional, variando de -1 (completamente opuestos) a 1 (idénticos). Cuanto mayor sea la similitud del coseno entre dos ítems, más se considerarán parecidos. Este enfoque permite determinar la semejanza entre los perfiles de los ítems y establecer conexiones significativas en el contexto de un sistema de recomendación.

similitudes = cosine_similarity(modelo_item.iloc[:,3:])

similitudes.shape

print(similitudes)

def recomendacion_juego(id):
    
    id = int(id)
    # Filtramos el juego e igualarlo a  su ID
    juego_seleccionado = modelo_item[modelo_item['item_id'] == id]
    # devolvemos error en caso de vacio
    if juego_seleccionado.empty:
        return "El juego con el ID especificado no existe en la base de datos."
    
    # Calculamos la matriz de similitud coseno
    #similitudes = cosine_similarity(modelo_item.iloc[:,3:])
    
    # Calculamos la similitud del juego que se ingresa con otros juegos del dataframe
    similarity_scores = similitudes[modelo_item[modelo_item['item_id'] == id].index[0]]
    
    # Calculamos los índices de los juegos más similares (excluyendo el juego de entrada)
    indices_juegos_similares = similarity_scores.argsort()[::-1][1:6]
    
    # Obtenemos los nombres de los juegos 5 recomendados
    juegos_recomendados = modelo_item.iloc[indices_juegos_similares]['app_name']
    
    return juegos_recomendados

recomendacion_juego(2028103)

Modelo para Railway

modelo_item

Dividimos el dataframe por la mitad a fin de lograr montar el modelo en render 

# Revisamos la cantidad de filas
cant_filas= len(modelo_item)

# Calculamos la mitad
mitad_filas= cant_filas // 10
 
# Seleccionamos la mitad superior
modelo_railway= modelo_item.iloc[:mitad_filas]

modelo_railway.shape

similitudes_railway = cosine_similarity(modelo_railway.iloc[:,3:])

Almacenamos el dataframe acotado 

modelo_railway.to_parquet("Datasets/modelo_railway.parquet")

Creamos una nueva función diseñada para explorar exclusivamente las similitudes con la fila del juego especificada como parámetro. Esto se realiza con el propósito de facilitar la ejecución del archivo mediante el proceso de renderizado.

def encontrar_juegos_similares(id):
    # Encontramos el índice del juego ingresado por ID
    juego_indice = modelo_railway.index[modelo_railway['item_id'] == id].tolist()[0]
    
    # Extraemos las características del juego ingresado
    juego_caracteristicas = modelo_railway.iloc[juego_indice, 3:].values.reshape(1, -1)
    
    # Calculamos la similitud coseno entre el juego ingresado y todos los demás juegos
    similitudes_railway = cosine_similarity(modelo_railway.iloc[:, 3:], juego_caracteristicas)
    
    # Obtenemos los índices de los juegos más similares (excluyendo el juego de entrada)
    indices_juegos_similares = similitudes_railway.argsort(axis=0)[::-1][1:6]
    indices_juegos_similares = indices_juegos_similares.flatten()[1:]
    
    # Obtenemos los juegos más similares en función de los índices
    juegos_similares = modelo_railway.iloc[indices_juegos_similares]['app_name']
    
    return juegos_similares

encontrar_juegos_similares(3310)

#### Creamos una función para obtener recomendaciones

def recomendacion_juego(id: int):
    
    # Verificamos si el juego con game_id existe en df_games
    game = modelo_render[modelo_railway['item_id'] == id]

    if game.empty:
        return("El juego '{id}' no posee registros.")
    
    # Obtenemos el índice del juego dado
    idx = game.index[0]

    # Tomamos una muestra aleatoria del DataFrame df_games
    sample_size = 2000  # Definimos el tamaño de la muestra (ajusta según sea necesario)
    df_sample = modelo_railway.sample(n=sample_size, random_state=42)  # Ajustamos la semilla aleatoria según sea necesario

    # Calculamos la similitud de contenido solo para el juego dado y la muestra
    sim_scores = cosine_similarity([modelo_railway.iloc[idx, 3:]], df_sample.iloc[:, 3:])

    # Obtenemos las puntuaciones de similitud del juego dado con otros juegos
    sim_scores = sim_scores[0]

    # Ordenamos los juegos por similitud en orden descendente
    similar_games = [(i, sim_scores[i]) for i in range(len(sim_scores)) if i != idx]
    similar_games = sorted(similar_games, key=lambda x: x[1], reverse=True)

    # Obtenemos los 5 juegos más similares
    similar_game_indices = [i[0] for i in similar_games[:5]]

    # Listamos los juegos similares (solo nombres)
    similar_game_names = df_sample['app_name'].iloc[similar_game_indices].tolist()

    return {"similar_games": similar_game_names}

recomendacion_juego(10)
